Separate card input classes to their own file.
