+++ sdk/src/main/java/com/checkout/sdk/cardinput/CardInput.kt
@@ -7,6 +7,7 @@ import android.text.InputFilter
 import android.text.SpannableStringBuilder
 import android.util.AttributeSet
 import android.view.View.OnFocusChangeListener
+import com.checkout.sdk.architecture.MvpView
 import com.checkout.sdk.architecture.PresenterStore
 import com.checkout.sdk.store.DataStore
 import com.checkout.sdk.utils.AfterTextChangedListener
@@ -24,7 +25,8 @@ import com.checkout.sdk.utils.CardUtils
 class CardInput @JvmOverloads constructor(
     internal var mContext: Context,
     attrs: AttributeSet? = null
-) : android.support.v7.widget.AppCompatEditText(mContext, attrs), CardInputPresenter.CardInputView {
+) : android.support.v7.widget.AppCompatEditText(mContext, attrs),
+    MvpView<CardInputUiState> {
 
 
     private var mCardInputListener: Listener? = null
@@ -66,7 +68,7 @@ class CardInput @JvmOverloads constructor(
         presenter.stop()
     }
 
-    override fun onStateUpdated(uiState: CardInputPresenter.CardInputUiState) {
+    override fun onStateUpdated(uiState: CardInputUiState) {
         // Get Card type
         filters =
                 arrayOf<InputFilter>(InputFilter.LengthFilter(uiState.cardType.maxCardLength))
@@ -93,7 +95,7 @@ class CardInput @JvmOverloads constructor(
         }
     }
 
-    private fun restoreCardNumberIfNecessary(cardInputResult: CardInputPresenter.CardInputUiState) {
+    private fun restoreCardNumberIfNecessary(cardInputResult: CardInputUiState) {
         if (text.isEmpty() && cardInputResult.cardNumber.isNotEmpty()) {
             setText(cardInputResult.cardNumber)
             setSelection(cardInputResult.cardNumber.length)
@@ -101,7 +103,7 @@ class CardInput @JvmOverloads constructor(
         }
     }
 
-    private fun showOrClearErrors(cardInputResult: CardInputPresenter.CardInputUiState) {
+    private fun showOrClearErrors(cardInputResult: CardInputUiState) {
         mCardInputListener?.let {
             if (cardInputResult.showCardError) {
                 it.onCardError()
