+++ sdk/src/main/java/com/checkout/sdk/cardinput/CardInputPresenter.kt
@@ -3,18 +3,13 @@ package com.checkout.sdk.cardinput
 import android.text.Editable
 import com.checkout.sdk.architecture.BasePresenter
 import com.checkout.sdk.architecture.MvpView
-import com.checkout.sdk.architecture.UiState
 import com.checkout.sdk.store.DataStore
-import com.checkout.sdk.utils.CardUtils
 
 
 class CardInputPresenter(
     private val dataStore: DataStore,
     initialState: CardInputUiState = CardInputUiState()
-) :
-    BasePresenter<CardInputPresenter.CardInputView, CardInputPresenter.CardInputUiState>(
-        initialState
-    ) {
+) : BasePresenter<MvpView<CardInputUiState>, CardInputUiState>(initialState) {
 
     fun textChanged(text: Editable) {
         val cardInputResult = CardInputUseCase(text, dataStore).execute()
@@ -34,14 +29,4 @@ class CardInputPresenter(
         )
         safeUpdateView(newState)
     }
-
-    data class CardInputUiState(
-        val cardNumber: String = "",
-        val cardType: CardUtils.Cards = CardUtils.Cards.DEFAULT,
-        val inputFinished: Boolean = false,
-        val showCardError: Boolean = false
-    ) : UiState
-
-    interface CardInputView : MvpView<CardInputUiState>
-
 }
