+++ sdk/src/main/java/com/checkout/sdk/cardinput/CardInputUiState.kt
@@ -1,0 +1,11 @@
+package com.checkout.sdk.cardinput
+
+import com.checkout.sdk.architecture.UiState
+import com.checkout.sdk.utils.CardUtils
+
+data class CardInputUiState(
+    val cardNumber: String = "",
+    val cardType: CardUtils.Cards = CardUtils.Cards.DEFAULT,
+    val inputFinished: Boolean = false,
+    val showCardError: Boolean = false
+) : UiState
